% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_addZ.R
\name{geojson_addZ}
\alias{geojson_addZ}
\title{Add Z value to geometries and export in a new file}
\usage{
geojson_addZ(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.out = NA,
  ids = "EAMENA ID",
  elevation.api = "gmrt_point",
  timeout = 30,
  sleep = 0.3,
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{path of GeoJSON file. By default 'caravanserail.geojson'.}

\item{geojson.out}{name of output file. By default NA, will add suffix 'Z' to name of GeoJSON file (variable `geojson.path`).}

\item{ids}{field having the unique keys, to remove duplicated geometries. By default "EAMENA ID".}

\item{elevation.api}{the geoserver API used to collect the elevation, by default 'gmrt_point' (https://www.gmrt.org). Another option is 'gmrt_profile', and 'open-elevation' (https://api.open-elevation.com/).}

\item{timeout}{the threshold below which the expectation of a response is accepted, in seconds. By default, 30.}

\item{sleep}{the time delay between two API request in seconds. By default 0.3.}

\item{verbose}{if TRUE (by default): verbose.}
}
\value{
An sf object with a Z column with Z values
}
\description{
Use geoserver API to recover the Z for given set of points stored in GeoJSON file. Geoserver API can lead to Timeout.
}
\examples{

df <- geojson_addZ()

# using 'open-elevation' as API
df <- geojson_addZ(elevation.api = 'open-elevation')


}
