% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_analysis.R
\name{geojson_analysis}
\alias{geojson_analysis}
\title{Complex interactive statistics on a GeoJSON file}
\usage{
geojson_analysis(
  stat.name = "stat",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  ids = ref_ids("hp.id"),
  export.stat = F,
  write.stat = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{stat.name}{the name of the output file. By default "stat".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{ids}{the IDs of the resources, by default "EAMENA.ID" (nb: R field names format, without spaces).}

\item{export.stat}{if TRUE return the stats to be stored in a new variable.}

\item{write.stat}{if TRUE, export the stats in a new file, if FALSE will only display it.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is write.stat is TRUE.}

\item{stat}{the statistic that will be computed. By default 'list_fields' (list the fields). The other options are: "list_ids" list EAMENA IDs ; etc.}
}
\value{
Show or export basic statistics on the GeoJSON file
}
\description{
Complex interactive statistics on a GeoJSON file with the Shiny package
}
\examples{

geojson_analysis(stat.name = "factorial analyis", export.stat = T)

}
