% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_hp_mds.R
\name{geojson_boxplot Create boxplot or various boxplots of path lengths between different heritage places}
\alias{geojson_boxplot Create boxplot or various boxplots of path lengths between different heritage places}
\alias{ref_hp_mds}
\title{Create boxplot or various boxplots of path lengths between different heritage places}
\usage{
ref_hp_mds(
  mds.template =
    "https://raw.githubusercontent.com/eamena-project/eamena-arches-dev/main/dbs/database.eamena/data/reference_data/rm/hp/mds/mds-template.xlsx",
  create.datatable = TRUE,
  mds.path =
    "C:/Rprojects/eamena-arches-dev/dbs/database.eamena/data/reference_data/rm/hp/mds/",
  mds.datatable.name = "fields-description.html",
  verbose = TRUE
)
}
\arguments{
\item{mds.template}{The XLSX file listing the HP fields, field descriptions, fields UUID in the DB, if these fields are part of the Minimum Data Standards (MDS), etc..}

\item{create.datatable}{If TRUE (default), will export an HTML datatable}

\item{mds.datatable.name}{The name of the HTML datatable}

\item{verbose}{if TRUE (Default), print messages.}
}
\value{
An HTML datatable (sortable, searchable, etc.)
}
\description{
Create boxplot or various boxplots of path lengths between different heritage places
}
