% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_boxplot.R
\name{geojson_boxplot}
\alias{geojson_boxplot}
\title{Create a boxplot or various boxplots of path lengths between different heritage places}
\usage{
geojson_boxplot(
  stat.name = "caravanserais_areas",
  concept.name = "hp.id",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  stat = c("area"),
  chart.type = c("boxplot"),
  by = NA,
  interactive = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 9,
  fig.height = 5,
  color.set = "Set1",
  verbose = TRUE
)
}
\arguments{
\item{stat.name}{the name of the output file and of the plot. By default "caravanserais_areas".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded.}

\item{stat}{the statistic that will be computed. The different values can be "area" for areas, or "dist" for distances between heritage places. By default 'area'.}

\item{chart.type}{the type of chart that will be plot. By default, "boxplot".}

\item{by}{the name of the field on which the paths will be grouped. For example "route". Will create as many plots as there is different categories. By default NA.}

\item{interactive}{if TRUE, create a Plotly chart. By default, FALSE.}

\item{export.plot}{if TRUE, will save the plot. By default, FALSE.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.If it doesn't exist, it will be created. Only useful is export plot is TRUE.}

\item{fig.width, fig.height}{size of the output chart.}

\item{color.set}{the RBrewer color set. By default "Set1".}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{

}
\description{

}
\examples{

# boxplot on areas (by default)
geojson_boxplot(stat = "area")

# boxplot on distances
geojson_boxplot(stat = "dist")

# Same as before, but stratified by routes, and saved
geojson_boxplot(stat.name = "caravanserais_areas", stat = "area", by = "route",
                export.plot = T)
geojson_boxplot(stat.name = "caravanserais_dist", stat = "dist", by = "route",
                export.plot = T)

# Same as before, but stratified by routes, interactive, and saved
geojson_boxplot(stat.name = "caravanserais_areas", stat = "area", by = "route",
                interactive = T,
                export.plot = T)
geojson_boxplot(stat.name = "caravanserais_dist", stat = "dist", by = "route",
                interactive = T,
                export.plot = T)

}
