% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_boxplot_path.R
\name{geojson_boxplot_path}
\alias{geojson_boxplot_path}
\title{Create a boxplot with paths distances between different heritage places}
\usage{
geojson_boxplot_path(
  plot.name = "box_path",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  plotly.plot = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 9,
  fig.height = 5
)
}
\arguments{
\item{plot.name}{the name of the output boxplot and the name of the saved file (if export.plot is TRUE). By default "box_path".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded.}

\item{plotly.plot}{if TRUE, export the plot, if FALSE will only display it}

\item{export.plot}{if TRUE, show the plot in Plotly window.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export plot is TRUE.}
}
\value{

}
\description{

}
\examples{

# normal plot
geojson_boxplot_path(export.plot = T)

# Plotly plot
geojson_boxplot_path(plotly.plot = T)

}
