% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_csv.R
\name{geojson_csv}
\alias{geojson_csv}
\title{Convert from GeoJSON to CSV}
\usage{
geojson_csv(
  geom.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_outGeoJSON.geojson"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  csv.name = "caravanserail_outCSV",
  verbose = T
)
}
\arguments{
\item{geom.path}{the path to the GeoJSON file. This file comes from the `geojson_kml()` function.}

\item{export}{if TRUE (by default), will export the GeoJSON to a CSV file.}

\item{dirOut}{the path to the folder where the CSV file will be created.}

\item{csv.name}{the name of the CSV that will be created without the extension}

\item{verbose}{if TRUE (by default) then display different messages.}
}
\value{
a CSV file
}
\description{
This function allows file format conversion from GeoJSON to CSV. The output CSV file then can be imported into the EAMENA DB with a bulk upload append procedure.
}
\examples{

geojson_csv()

}
