% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_get_field.R
\name{geojson_get_field}
\alias{geojson_get_field}
\title{Get values of a given field in a GeoJSON file.}
\usage{
geojson_get_field(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  field = "EAMENA ID"
)
}
\arguments{
\item{geojson.path}{the path to the GeoJSON file.}

\item{field}{a field name, by default "EAMENA ID".}
}
\value{
A vector with all values
}
\description{
Get all values of a given field.
}
\examples{

geojson_get_field()
##  [1] "EAMENA-0192281" "EAMENA-0182033" "EAMENA-0182054" "EAMENA-0182055" "EAMENA-0182056" "EAMENA-0182057" "EAMENA-0182058"
##  [8] "EAMENA-0164899" "EAMENA-0164904" "EAMENA-0164906" "EAMENA-0164905" "EAMENA-0164943" "EAMENA-0164999" "EAMENA-0207261"
## ...

}
