% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_geojson_map.R, R/_geojson_map_1.R,
%   R/_geojson_map_2.R, R/geojson_map.R
\name{geojson_map}
\alias{geojson_map}
\title{Create map, whether static or interactive leaflet, from GeoJSON file}
\usage{
geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.grids = NA,
  zoom = NA,
  ids = "EAMENA ID",
  field.names = NA,
  symbology = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  hp.color = "red",
  hp.color.bck = "grey",
  hp.size = 2,
  maptype = c("esri", "world_imagery"),
  fields.for.labels = c("Site Feature Interpretation Type", "Cultural Period Type",
    "Administrative Division ", "Country Type ", "Overall Condition State Type"),
  ncol = 4,
  max.maps = NA,
  interactive = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)

geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.grids = NA,
  zoom = NA,
  ids = "EAMENA ID",
  field.names = NA,
  symbology = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  hp.color = "red",
  hp.color.bck = "grey",
  hp.size = 2,
  maptype = c("esri", "world_imagery"),
  fields.for.labels = c("Site Feature Interpretation Type", "Cultural Period Type",
    "Administrative Division ", "Country Type ", "Overall Condition State Type"),
  ncol = 4,
  max.maps = NA,
  interactive = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)

geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.grids = NA,
  zoom = NA,
  ids = "EAMENA ID",
  field.names = NA,
  symbology = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  hp.color = "red",
  hp.color.bck = "grey",
  hp.size = 2,
  maptype = c("esri", "world_imagery"),
  fields.for.labels = c("Site Feature Interpretation Type", "Cultural Period Type",
    "Administrative Division ", "Country Type ", "Overall Condition State Type"),
  ncol = 4,
  max.maps = NA,
  interactive = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)

geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.grids = NA,
  zoom = NA,
  ids = "EAMENA ID",
  field.names = NA,
  symbology = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  hp.color = "red",
  hp.color.bck = "grey",
  hp.size = 2,
  maptype = c("esri", "world_imagery"),
  fields.for.labels = c("Site Feature Interpretation Type", "Cultural Period Type",
    "Administrative Division ", "Country Type ", "Overall Condition State Type"),
  ncol = 4,
  max.maps = NA,
  interactive = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)
}
\arguments{
\item{map.name}{name of output map and name of saved file (if export.plot is TRUE). Default "map".}

\item{geojson.path}{path of GeoJSON file. Default 'caravanserail.geojson'.}

\item{geojson.grids}{optional. path of GeoJSON file to show also grids.}

\item{zoom}{An integer value for the zoom for the Stadiamap basemap. Default NA, will be calculated automatically.}

\item{ids}{IDs of resources, default "EAMENA.ID" (R fieldname format, without spaces).}

\item{field.names}{a vector of one or many field names for thematic cartography. If NA (default), will create a general map}

\item{symbology}{path to XLSX recording symbology for different values, default 'symbology.xlsx'.}

\item{maptype}{type of background basemap, from the  basemaps:: package}

\item{ncol}{the number of columns}

\item{max.maps}{to limit the total number of exported maps. Default: NA.}

\item{interactive}{if FALSE create static PNG (by  default), if TRUE create a plotly plot as HTML widget.}

\item{export.plot}{if TRUE, export plot, if FALSE will only display it.}

\item{dirOut}{folder where outputs will be saved. Default: '/results'. If it doesn't exist, will be created. Only useful if export plot is TRUE.}

\item{fig.width, fig.height}{size of output map.}

\item{verbose}{if TRUE (default) then display different messages.}

\item{stamen.zoom}{zoom of Stamen basemap, between 0 (world, unprecise) to 21 (building, very precise). Default NA, zoom level will be calculated automatically.}

\item{highlights.ids}{EAMENA IDs (ex: 'EAMENA-0205783') that will be highlighted in map. If NA (default), no highlights.}
}
\value{
An interactive map (leaflet) or not

An interactive map (leaflet) or not

An interactive map (leaflet) or not

An interactive map (leaflet) or not
}
\description{
Create distribution map

Create distribution map

Create distribution map

Create distribution map
}
\examples{


# plot a general map of heritage places
geojson_map(map.name = "caravanserail")

# save a thematic map
geojson_map(map.name = "caravanserail",
           field.names = c("Damage Extent Type"),
           export.plot = T)

# save different thematic maps
geojson_map(map.name = "caravanserail",
           field.names = c("Disturbance Cause Type ", "Damage Extent Type"),
           export.plot = T)

# save an interactive map
geojson_map(map.name = "caravanserail_plotly",
            interactive = T,
            export.plot = T)

# plot a general map of geoarchaeological data
geojson_map(map.name = "geoarch",
            ids = "GEOARCH.ID",
            stamen.zoom = 6,
            geojson.path = "C:/Rprojects/eamena-arches-dev/data/geojson/geoarchaeo.geojson",
            export.plot = F)



# plot a general map of heritage places
geojson_map(map.name = "caravanserail")

# save a thematic map
geojson_map(map.name = "caravanserail",
           field.names = c("Damage Extent Type"),
           export.plot = T)

# save different thematic maps
geojson_map(map.name = "caravanserail",
           field.names = c("Disturbance Cause Type ", "Damage Extent Type"),
           export.plot = T)

# save an interactive map
geojson_map(map.name = "caravanserail_plotly",
            interactive = T,
            export.plot = T)

# plot a general map of geoarchaeological data
geojson_map(map.name = "geoarch",
            ids = "GEOARCH.ID",
            stamen.zoom = 6,
            geojson.path = "C:/Rprojects/eamena-arches-dev/data/geojson/geoarchaeo.geojson",
            export.plot = F)



# plot a general map of heritage places
geojson_map(map.name = "caravanserail")

# save a thematic map
geojson_map(map.name = "caravanserail",
           field.names = c("Damage Extent Type"),
           export.plot = T)

# save different thematic maps
geojson_map(map.name = "caravanserail",
           field.names = c("Disturbance Cause Type ", "Damage Extent Type"),
           export.plot = T)

# save an interactive map
geojson_map(map.name = "caravanserail_plotly",
            interactive = T,
            export.plot = T)

# plot a general map of geoarchaeological data
geojson_map(map.name = "geoarch",
            ids = "GEOARCH.ID",
            stamen.zoom = 6,
            geojson.path = "C:/Rprojects/eamena-arches-dev/data/geojson/geoarchaeo.geojson",
            export.plot = F)



# plot a general map of heritage places
geojson_map(map.name = "caravanserail")

# save a thematic map
geojson_map(map.name = "caravanserail",
           field.names = c("Damage Extent Type"),
           export.plot = T)

# save different thematic maps
geojson_map(map.name = "caravanserail",
           field.names = c("Disturbance Cause Type ", "Damage Extent Type"),
           export.plot = T)

# read from Zenodo and return a list of maps
map.name <- "Sistan_dataset_Damages_Extent"
all.g <- geojson_map(map.name = map.name,
                     field.names = c("Damage Extent Type"),
                     geojson.path = "https://doi.org/10.5281/zenodo.10375902",
                     max.maps = 6,
                     hp.color = "black",
                     hp.color.bck = "grey",
                     hp.size = 1.5)

# save an interactive map
geojson_map(map.name = "caravanserail_plotly",
            interactive = T,
            export.plot = T)

# plot a general map of geoarchaeological data
geojson_map(map.name = "geoarch",
            ids = "GEOARCH.ID",
            geojson.path = "C:/Rprojects/eamena-arches-dev/data/geojson/geoarchaeo.geojson",
            export.plot = F)

}
