% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_geojson_map_path.R, R/geojson_map_path.R
\name{geojson_map_path}
\alias{geojson_map_path}
\title{Create map and profile with paths between different heritage places (HP)}
\usage{
geojson_map_path(
  d = NA,
  map.name = "paths",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"), "/extdata/caravanserail_paths.csv"),
  stats = c("map"),
  by = "route",
  show.ids = TRUE,
  selected.category = NA,
  symbology.path = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  interactive = FALSE,
  stamen.zoom = NA,
  color.set = "Set1",
  verbose = TRUE
)

geojson_map_path(
  d = NA,
  map.name = "paths",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"), "/extdata/caravanserail_paths.csv"),
  stats = c("map"),
  by = "route",
  show.ids = TRUE,
  selected.category = NA,
  symbology.path = paste0(system.file(package = "eamenaR"), "/extdata/symbology.xlsx"),
  interactive = FALSE,
  stamen.zoom = NA,
  color.set = "Set1",
  verbose = TRUE
)
}
\arguments{
\item{d}{A hash dictionary.}

\item{map.name}{name of output map and name of saved file (if export.plot is TRUE). Default "map_path".}

\item{geojson.path}{path of GeoJSON file. Default 'caravanserail.geojson'.}

\item{csv.path}{path to CSV where edges between two HPs are recorded. Default 'caravanserail_paths.csv'.}

\item{stats}{type of output: map (`map`) or profile (`profile`). For latter Z should be calculated with the `geojson_addZ()` function.}

\item{show.ids}{Show HP labels. Default: TRUE.}

\item{selected.category}{limit study to some categories. For example to some particular routes for caravanserais. Default NA, no limitation.}

\item{symbology.path}{path to XLSX recording symbology for different values, default 'symbology.xlsx'.}

\item{interactive}{if TRUE will plot VisNetwork. Default FALSE.}

\item{stamen.zoom}{zoom of Stamen basemap, between 0 (world, unprecise) to 21 (building, very precise). By default NA, zoom level will be calculated automatically.}

\item{color.set}{RBrewer color set. Default "Set1".}

\item{verbose}{if TRUE (by default), print messages.}

\item{export.type}{type of output: map (`map`) or profile (`profile`). For latter Z should be calculated with the `geojson_addZ()` function.}

\item{symbology}{path to XLSX recording symbology for different values, default 'symbology.xlsx'.}

\item{export.plot}{if TRUE, export plot, if FALSE will only display it.}

\item{dirOut}{folder where outputs will be saved. Default: '/results'. If it doesn't exist, will be created. Only useful if export plot is TRUE.}

\item{fig.width, fig.height}{size of output map.}

\item{name}{of category column. Default "route" for caravanserais.}
}
\value{
PNG map and HMTL map of HPs linked together by paths

An hash object with maps of HPs linked together by paths, profiles, etc.
}
\description{
Create distribution map and elevation profile of HP linked together by paths, for example, caravanserails.

Create distribution map and elevation profile of HP linked together by paths, for example, caravanserails.
}
\examples{

library(dplyr)

# plot general map of HPs
geojson_map_path(map.name = "caravanserail_paths", export.plot = F)

# plot and export
geojson_map_path(map.name = "caravanserail_paths", export.plot = T, fig.width = 11)

# create an interactive map of each route
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T)

# create interactive map of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T,
                 export.plot = T,
                 dirOut = "C:/Rprojects/eamenaR/results/")

# create profile of each route
df <- geojson_addZ(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
                   dirOut = "C:/Rprojects/eamenaR/inst/extdata/")
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 export.type = "profile")

# create profile of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 export.type = "profile",
                 export.plot = T,
                 fig.height = 11,
                 fig.width = 18,
                 dirOut = "C:/Rprojects/eamenaR/results/")


library(dplyr)

# plot general map of HPs
d <- hash::hash()
d <- geojson_map_path(map.name = "caravanserail_paths")
d$caravanserail_paths_map


# create an interactive map of each route
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T)

# create interactive map of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T,
                 export.plot = T,
                 dirOut = "C:/Rprojects/eamenaR/results/")

# create profile of each route
df <- geojson_addZ(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
                   dirOut = "C:/Rprojects/eamenaR/inst/extdata/")
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 stats = "profile")

# create profile of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 stats = "profile",
                 export.plot = T,
                 fig.height = 11,
                 fig.width = 18,
                 dirOut = "C:/Rprojects/eamenaR/results/")

}
