% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_measurements.R
\name{geojson_measurements}
\alias{geojson_measurements}
\title{Measurements on a GeoJSON file}
\usage{
geojson_measurements(
  stat.name = "stat",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  stat = c("area"),
  by.category = NA,
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  plot.stat = TRUE,
  export.stat = FALSE,
  export.plot = FALSE,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)
}
\arguments{
\item{stat.name}{the name of the output file. By default "stat".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{stat}{the statistic that will be computed. By default 'area'.}

\item{by.category}{the name of the field that will be used to stratified the values. For example "route" for the caravanserais. By default NA (no stratification).}

\item{csv.path}{if the parameter by.category is TRUE, will use this CSV file to recover the route of the heritage places}

\item{plot.stat}{if TRUE (by default) will plot the stat as a graphic.}

\item{export.stat}{if TRUE return the stats to be stored in a new variable.}

\item{export.plot}{if TRUE, export the stats in a new file.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export.plot is TRUE.}

\item{verbose}{if TRUE (by default) verbose.}
}
\value{
Show or export measurements statistics on the GeoJSON file
}
\description{
Compute measurements (areas, L x l, etc.) on a GeoJSON file and output a boxplot
}
\examples{

# boxplot for all heritage places
geojson_measurements(stat.name = "areas")

# by route and export
geojson_measurements(stat.name = "areas", by.category = "route", export.stat = T)

}
