% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_read.R
\name{geojson_read}
\alias{geojson_read}
\title{Read a GeoJSON file}
\usage{
geojson_read(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{path of GeoJSON file. Default 'caravanserail.geojson'.}
}
\value{
A list with a sf object (the GeoJSON file) and a caption
}
\description{
Read a GeoJSON file from a local path or a Zenodo deposit
}
