% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_shp.R
\name{geojson_shp}
\alias{geojson_shp}
\title{Convert between GeoJSON and SHP}
\usage{
geojson_shp(
  geom.path = paste0(system.file(package = "eamenaR"), "/extdata/Waypoints.kmz"),
  geom.types = c("POINT", "LINE", "POLYGON"),
  export = TRUE,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  geojson.name = "Waypoints",
  select.name = "EAMENA ID",
  select.fields = c("EAMENA ID", "Resource Name", "resourceid"),
  verbose = TRUE
)
}
\arguments{
\item{geom.path}{path to the SHP/GeoJSONfile.}

\item{geom.types}{types of geometries ("POINT", "LINE", "POLYGON") that will be selected. Default all: `c("POINT", "LINE", "POLYGON")`.}

\item{export}{if TRUE, will export SHP file in GeoJSON format, or GeoJSON file as SHP, if FALSE simple plot.}

\item{dirOut}{path to folder where SHP/GeoJSON file will be created. Default, 'results/'.}

\item{geojson.name}{name of SHP/GeoJSON that will be created without extension}

\item{select.name}{for SHP export only. Field selected to be SHP name of the heritage place, default `"EAMENA ID"`.}

\item{select.fields}{for SHP export only. SHP conversion remove large number of GeoJSON fields. This variable is used to select fields to preserve. Default: `c("EAMENA ID","Resource Name", "resourceid")`.}

\item{verbose}{if TRUE (default) then display different messages.}
}
\value{
GeoJSON file or SHP file, depending on input file.
}
\description{
Convert from SHP (shapefile) to GeoJSON, or from GeoJSON to SHP.
}
\examples{

library(dplyr)

# GeoJSON to SHP, only POINTS
geojson_shp(geom.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
            geom.types = "POINT",
            dirOut = "C:/Rprojects/eamenaR/results/",
            export = T,
            geojson.name = "caravanserailOUT")


# GeoJSON to SHP, only POLYGONS
library(dplyr)

geojson_shp(geom.types = "POLYGON",
           geojson.name = "cvns-polygons",
           geom.path = paste0(system.file(package = "eamenaR"),
                              "/extdata/caravanserail.geojson"),
           dirOut = "C:/Rprojects/eamena-gee/data/")

}
