% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bbox.R
\name{geom_bbox}
\alias{geom_bbox}
\title{Create the bounding box of BUs' geometries and save it as a GeoJSON file}
\usage{
geom_bbox(
  dataDir = NA,
  x_column = NA,
  y_column = NA,
  wkt_column = NA,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  geojson.name = "mbr.geojson",
  verbose = TRUE
)
}
\arguments{
\item{dataDir}{the path to the folder where the XLSXs are.}

\item{x_column, y_column}{the column of the X and Y coordinates, if NA the function will look the field 'wkt'.}

\item{wkt_column}{the column of the WKT coordinates, if NA (by default) the function will look the field 'x_column' and 'y_column'.}

\item{dirOut}{the folder where the GeoJSON will be saved. By default: '/results'. If it doesn't exist, it will be created.}

\item{geojson.name}{the name of the GeoJSON that will be created, by default 'mbr' (minimum bound rectangle)}

\item{verbose}{if TRUE (by default) then display different messages.}
}
\value{
the bounding box of the dataset as a GeoJSON file
}
\description{
Read one or various XLSX worksheets within a folder. Collect the xmin, xmax, ymin, xmax in the coordinates and creates the minimum bounding box (MBR), that is to say, the extent of the geometries. The export is a GeoJSON file that can be copy/paste in the Map filter of EAMENA database to recover the selected grid cells(\link[eamenaR]{list_mapping_bu.R}). Can also be used to get the precise Stamen basemap extent (ggmap), etc.
}
\examples{

geom_bbox(dataDir = "C:/Rprojects/eamena-arches-dev/data/bulk/bu/mk2")

geom_bbox(dataDir = "C:/Rprojects/eamena-arches-dev/data/bulk/bu/mk2/",
          x_column = "Longitude.",
          y_column = "Latitude",
          geojson.name = "grids_bbox_mk2.geojson")

}
