% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kml.R
\name{geom_kml}
\alias{geom_kml}
\title{Extracting KMZ or KML coordinates from a file and adding this coordinates into EAMENA for a same HP}
\usage{
geom_kml(
  kml.path = paste0(system.file(package = "eamenaR"), "/extdata/Waypoints.kmz"),
  geom.types = c("POINT", "LINE", "POLYGON"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  geojson.name = "Waypoints.geojson",
  verbose = T
)
}
\arguments{
\item{kml.path}{the path to the KML or KMZ file.}

\item{geom.types}{the types of geometries ("POINT", "LINE" or "POLYGON") that will be selected. By default all: `c("POINT", "LINE", "POLYGON")`.}

\item{export}{if TRUE, will export the KML/KMZ file in a GeoJSON format.}

\item{dirOut}{the path to the folder where the GeoJSON file will be created.}

\item{geojson.name}{the name of the GeoJSON that will be created.}

\item{verbose}{if TRUE (by default) then display different messages.}
}
\value{
a GeoJSON file
}
\description{
Geometries drawn in Google Earth or Google Earth pro can be exported as KML or KMZ (ie, compressed KML). For a given HP, this data can be then imported into the EAMENA DB.
}
\examples{

geom_kml(kml.path = "C:/Rprojects/eamenaR/inst/extdata/Waypoints.kmz",
         dirOut = "C:/Rprojects/eamenaR/inst/extdata/",
         geom.types = "POLYGON",
         export = T,
         geojson.name = "Waypoints.geojson")



}
