% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_within_gs.R
\name{geom_within_gs}
\alias{geom_within_gs}
\title{Test if a resource (HP) geometry is within a Grid Square.}
\usage{
geom_within_gs(resource.wkt = NA, grid.squares.sf = NA)
}
\arguments{
\item{resource.wkt}{the WKT geometry of a resource, as a character format. This
WKT geometry can comes from a BU sheet (ex: POINT(0.916350216921341 35.9625191284127))}

\item{grid_squares}{a path to a GeoJSON file. This GeoJSON is the
EAMENA output of the GeoJSON URL. The GeoJSON is read and comvert to a sf object}
}
\value{
the ID of the Grid Square
}
\description{
Test if the geometry of a resource (eg. Heritage Place) is
within a Grid Square (gs). If so, return the ID of the Grid Square
}
\examples{

}
