% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_concepts.R
\name{list_concepts}
\alias{list_concepts}
\title{List the name of all the child-concepts below a certain Concept node}
\usage{
list_concepts(db.con = NA, d = NA, field = NA, uuid = NA)
}
\arguments{
\item{db.con}{a `dbConnect` connection to the database}

\item{d}{a hash() object (a Python-like dictionary)}

\item{field}{the field name that will be created in the a hash() object}

\item{uuid}{the UUID of the Concept parent. For example, '3b5c9ac7-5615-3de6-9e2d-4cd7ef7460e4' is the UUID of ..}
}
\value{
A dataframe stored in the input hash() object, under the selected 'field' name. This dataframe will with listed child-concepts in the provided field name. The UUID of each sub-concept will be stored into the 'field.uuid' column of the dataframe
}
\description{
With a given concept UUID (v. Reference Data Manager), find all
the childs. This function has been tested only for Cultural and Subcultural periods
}
\examples{
.

# in the connection parameters, replace 'xxx' by the password
d <- hash::hash()
my_con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                    user = 'postgres',
                    password = xxx,
                    dbname = 'eamena',
                    host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                    port = 5432)
# Cultural periods
d <- list_concepts(db.con = my_con, d = d, field = "cultural_periods", uuid = '3b5c9ac7-5615-3de6-9e2d-4cd7ef7460e4')

# Subcultural periods
d <- list_concepts(db.con = my_con, d = d, field = "subcultural_periods", uuid = '16cb160e-7b31-4872-b2ca-6305ad311011')

# disconnect
RPostgres::dbDisconnect(my_con)

}
