% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_culturalper.R
\name{list_culturalper}
\alias{list_culturalper}
\title{List the name of all the cultural period of a given HP}
\usage{
list_culturalper(
  db = "eamena",
  d,
  field,
  uuid = NA,
  geojson.path = NA,
  raw.GH = "https://raw.githubusercontent.com/eamena-oxford/eamena-arches-dev/main/"
)
}
\arguments{
\item{db}{the name of the database or dataset, by default 'eamena'.
If 'eamena': will connect the Pg database.
If 'geojson', will read the GeoJSON file path
recorded in the parameter 'geojson.path'}

\item{d}{a hash() object (a Python-like dictionary)}

\item{field}{the field name that will be created in the a hash() object}

\item{uuid}{the UUID of the HP, only useful if db = 'eamena'}

\item{geojson.path}{the path of the GeoJSON file}
}
\value{
A hash() with listed cultural periods names
}
\description{
With a given concept UUID (v. Reference Data Manager), find all
the cultural periods, subperiods, etc., of a given HP
}
\examples{

d_sql <- hash::hash()
d_sql <- uuid_from_eamenaid("eamena", "EAMENA-0187363", d_sql, "uuid")
d_sql <- list_culturalper("eamena", d_sql, "culturalper", d_sql$uuid)

}
