% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mapping_bu_append.R
\name{list_mapping_bu_append}
\alias{list_mapping_bu_append}
\title{Convert different types of files (GeoJSON, CSV, etc.) into an appropriate structure for a Bulk Upload append procedure}
\usage{
list_mapping_bu_append(
  fileIn = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_outGeoJSON.geojson"),
  method = "geom",
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  csv.name = "caravanserail_outCSV",
  verbose = T
)
}
\arguments{
\item{fileIn}{path to the file. Default: path to a GeoJSON file which comes from the `geojson_kml()` function.}

\item{method}{the type of transformation that will be done. Default `geom`: Convert from GeoJSON to CSV. Other transformations: `ir` for Information Resources.}

\item{export}{if TRUE (default), will export GeoJSON to CSV file.}

\item{dirOut}{path to folder where CSV file will be created.}

\item{csv.name}{name of CSV that will be created without the extension}

\item{verbose}{if TRUE (default) then display different messages.}
}
\value{
a CSV file
}
\description{
Function allows file format conversion from GeoJSON to CSV. Output CSV file can be imported into EAMENA DB with bulk upload append procedure.
}
\examples{

## Geometries
list_mapping_bu_append()

## List of related IR (HP <-> IR)


}
