% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodo_cultural_periods.R
\name{periodo_cultural_periods}
\alias{periodo_cultural_periods}
\title{Convert PeriodO CSV into an interactive data.table}
\usage{
periodo_cultural_periods(
  data.name = "periodo",
  periodo.url = "https://n2t.net/ark:/99152/p0dataset.csv",
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{data.name}{the name of the data.table. Only useful if export.plot is TRUE.}

\item{periodo.url}{the URL of the PeriodO CSV.}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export plot is TRUE.}
}
\value{
A DT data.frame of the Periodo cultural periods.
}
\description{
read a release of the periodo cultural periods thesaurus and creates an interactive dataframe.
}
\examples{


}
