% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cultural_periods.R
\name{plot_cultural_periods}
\alias{plot_cultural_periods}
\title{Plot the duration of EAMENA HP cultural periods attribution in a chart. The cultural periods
are recorded by years}
\usage{
plot_cultural_periods(
  d = NA,
  field = NA,
  type.plot = "static",
  bin.width = 50,
  export.plot = F,
  dataDir = paste0(system.file(package = "eamenaR"), "/results/"),
  seg.size = 1,
  seg.alpha = 0.5,
  fig.width = 8,
  fig.height = 8
)
}
\arguments{
\item{d}{a hash() object (a Python-like dictionary)}

\item{field}{the field name where the periods, subperiods, etc. will be read in the a hash() object
to plot it. It should be 'periods' or 'subperiods'}

\item{type.plot}{if "static" create a PNG, if "dynamic" create a HTML widget}

\item{bin.width}{size of the bins, by default, 50 years}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it}

\item{dataDir}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created.}

\item{seg.size, seg.alpha, fig.width, fig.height}{parameters for the gglpot output}
}
\value{
A plotly chart if 'dynamic', or a gglpot if 'static', to display or save
}
\description{
Read the 'cultural_periods.tsv' hosted on GitHub to find the tpq and taq dates
}
\examples{

d_sql <- hash::hash()
d_sql <- uuid_from_eamenaid("eamena", "EAMENA-0187363", d_sql, "uuid")
d_sql <- list_cultural_periods(db = "eamena", d = d_sql, field = "culturalper", uuid = d_sql[["uuid"]])
plot_cultural_periods(d = d_sql, field = "culturalper", export.plot = TRUE)

}
