% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_edtf.R
\name{plot_edtf}
\alias{plot_edtf}
\title{Create an aoristic analysis of cultural heritage threats from a dataset.}
\usage{
plot_edtf(
  data_file = paste0(system.file(package = "eamenaR"),
    "/extdata/disturbances_edtf.xlsx"),
  date_column = "EDTF",
  site_column = "S_ID",
  cause_column = "Disturbance.Cause",
  type_column = "Disturbance.Type",
  edft_limits = "2004-01-01..2019-12-31",
  edtf_span = "myd",
  edtf_analyse = c("all", "category"),
  edtf_analyse_category = "Disturbance.Type",
  edft_round = 4,
  id.filter = NA,
  type.plot = c("plotly"),
  export.plot = FALSE,
  file_out = "df_syria_out2",
  dirOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{data_file}{the path to the dataset. By default "disturbances_edtf.xlsx".}

\item{date_column}{the column of the EDTF data.}

\item{site_column}{the column of the sites' names or sites' IDs.}

\item{cause_column}{the column of the threats causes.}

\item{type_column}{the column of the type of threats.}

\item{edtf_span}{the time unit of analysis: "myd" = years, months and days (default); "my" = years and months; "y" = year.}

\item{edtf_analyse}{type of analysis. If "all" will sum the different categories of threats, if "category" will plot all categories of threats separately, on a same graphic. By default both: `edtf_analyse = c("all", "category")`}

\item{edtf_analyse_category}{the field name of the threat category that will be analysed: "Disturbance.Type" (default), or "Disturbance.Cause".}

\item{edft_round}{round precision for density, default 4 decimal places.}

\item{id.filter}{if not NA, will filter on this subset of sites (e.g. `id.filter = c("AM009")`)}

\item{type.plot}{file extension of the plot to be exported. Either "plotly" for a dynamic HTML plot, or "png" for a static plot.}

\item{export.plot}{if TRUE, saves the plot, if FALSE (default), displays the plot.}

\item{file_out}{the name of the plot that will be created if `export.plot` is TRUE}

\item{edtf_limits}{the min and max dates of the time analysis. This interval will be crossed with the boundaries of the dataset. Useful when threats are recorded "before" (end date) or "after" (start date) a specific date.}

\item{dataOut}{the path to the folder where the plot file will be created if `export.plot` is TRUE}
}
\value{
One to several plots
}
\description{
Use an XLSX file where cultural threats are listed and dated with EDTF.
R doesn't accept spaces in headers, so the spaces are replaced by dots.
}
\examples{

library(dplyr)
plot_edtf(data_file = "C:/Rprojects/eamena-arches-dev/data/time/Disturbances_EDTF.xlsx")

}
