% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hps.R
\name{plot_hps}
\alias{plot_hps}
\title{Plot the map distribution of Heritage Places created by year}
\usage{
plot_hps(
  df = NA,
  stadia_map_token = "aa5c9739-90c7-410b-9e9b-6c904df6e4dd",
  buff = 2,
  verbose = TRUE
)
}
\arguments{
\item{df}{A dataframe created with the `ref_hps()` function}

\item{stadia_map_token}{A StadiaMap token for the maps background}
}
\value{
A list of ggplot
}
\description{
Plot the map distribution of Heritage Places created by year
}
