% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_cultural_periods.R
\name{ref_cultural_periods}
\alias{ref_cultural_periods}
\title{Create a list of child-concepts below Cultural Period of all periods with their duration}
\usage{
ref_cultural_periods(db.con = NA, d = NA, field = NA)
}
\value{
NA
}
\description{
create a list concepts below Cultural Period of all periods
with their durations. Duration of each period are listed in the 'scopeNote' of this period. A periodo colum is added to the output dataframe. If 'export.table' then write a CSV file
}
\examples{

d <- hash::hash()
my_con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                    user = 'postgres',
                    password = xxx,
                    dbname = 'eamena',
                    host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                    port = 5432)

# get cultural periods
d <- list_concepts(db.con = my_con, d = d, field = "cultural_periods", uuid = '3b5c9ac7-5615-3de6-9e2d-4cd7ef7460e4')
d <- ref_cultural_periods(db.con = my_con, d = d, field = "cultural_periods")

# get subcultural periods
d <- list_concepts(db.con = my_con, d = d, field = "subcultural_periods", uuid = '16cb160e-7b31-4872-b2ca-6305ad311011')
d <- ref_cultural_periods(db.con = my_con, d = d, field = "subcultural_periods")

# and export as TSV
df.periods <- rbind(d$cultural_periods, d$subcultural_periods)
tout <- paste0(paste0(system.file(package = "eamenaR"), "/extdata/"), "cultural_periods.tsv")
write.table(df.periods, tout, sep ="\t", row.names = F)

# also, remember to disconnect
RPostgres::dbDisconnect(my_con)

}
