% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_hps.R
\name{ref_data}
\alias{ref_data}
\alias{ref_hps}
\title{Statistics about EAMENA business data: Heritage places (spatial distribution, nb of HP by grids, etc.), Grids, etc.}
\usage{
ref_hps(
  db.con = NA,
  d = NA,
  stat = c("spat"),
  perc = FALSE,
  rounded = 1,
  stat.name = "eamena_hps",
  stat.field = "Overall Condition State Type",
  stat.format = ".geojson",
  plot.map = FALSE,
  export.plot = FALSE,
  fig.width = 6,
  fig.height = 6,
  fig.dev = "png",
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  date.after = NA,
  date.before = NA,
  on.date = "assess.activity.date",
  max.num = NA,
  team.name = NA,
  verbose = TRUE
)
}
\arguments{
\item{db.con}{the parameters for the Postgres EAMENA DB, in a `RPostgres::dbConnect()` format.}

\item{d}{a hash() object (a Python-like dictionary).}

\item{stat}{the type of statistic that will be computed. Default: "spat" (spatial).Possible values: "grid" for stats on grids; "hist" for stats on categories having a large number of possible values (ex: Disturbances); "pie" for stats on categories having a small number of possible values (ex: Condition).}

\item{perc}{Percentage (if TRUE) or counts (default).}

\item{stat.name}{the name of the output file. By default "eamena_hps".}

\item{stat.field}{a `db.concept.name` listed in the 'ids.csv' file to use for the stats, for example 'Disturbance Cause Category Type', 'Overall Condition State Type', etc.}

\item{stat.format}{the extension of the geographic file (".geojson", ".shp"). GeoJSON by default.}

\item{plot.map}{if TRUE will plot a map (FALSE by default).}

\item{export.plot}{if TRUE will export the plot (FALSE by default).}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export.plot is TRUE.}

\item{date.after}{the date after which the calculation is made. Useful to limit the analysis. By default, NA.}

\item{date.before}{the date before which the calculation is made. Useful to limit the analysis. By default, the current date (`Sys.Date()`)}

\item{on.date}{whether to limit on "createdate" (a Postgres inner field) or on the field "Assessment Activity Date" (default)}

\item{max.num}{The maximum number of values (eg categories) to display, for example in an histogram. Default: NA.}

\item{team.name}{only the HPs from this team. Useful to limit the analysis. By default, NA (all the teams). For examples, the possible values to limit the analysis for the EAMENA DB are: "EAMENA Project Staff", "MarEA Project Staff", "Government Authority/Staff", "Volunteer/Independent Researcher", "Student/Trainee", "Academic Researcher", "Private Sector", "Non-Governmental Organisation (NGO)", etc.}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{
a hash() object. If plot.map and export.plot are set to TRUE will also create and save maps (SHP or GeoJSON).
}
\description{
statistics about EAMENA Heritage places. For example the HPs created in 2022, number of HP by grids, type of disturbances, overall condition state type of heritage places. This function is the backend counterpart of `geojson_stat()` for GeoJSON files.
}
\examples{

# Hash dictionary and Postgres connection (change the 'xxx' with usernames and password)

d <- hash::hash()
db.con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                               user = 'xxx',
                               password = 'xxx',
                               dbname = 'eamena',
                               host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                               port = 5432)

# Heritage places created during the year 2022 as a SHP file
d <- ref_hps(db.con = db.con,
            d = d,
            date.after = '2021-12-31',
            date.before = '2023-01-01',
            stat.name = "eamena_hps_2022",
            stat.format = ".shp",
            dirOut = 'C:/Rprojects/eamena-arches-dev/data/geojson/',
            export.plot = TRUE)

# Heritage places created during the year 2023 as a GeoJSON file
d <- ref_hps(db.con = db.con,
             d = d,
             date.after = '2022-12-31',
             date.before = '2024-01-01',
             stat.name = "eamena_hps_2023",
             stat.format = ".geojson",
             team.name = "EAMENA Project Staff",
             dirOut = 'C:/Rprojects/eamena-arches-dev/data/geojson/',
             export.plot = TRUE)

# Number of HP by grids
d <- hash::hash()
d <- ref_hps(db.con = db.con,
             d = d,
             stat = "grid",
             stat.name = "eamena_hps_by_grids",
             date.before = NA)

# histogram on Disturbance Cause Category Type
d <- ref_hps(db.con = db.con,
             d = d,
             stat = "hist",
             stat.name = "Disturbance Cause Category Type",
             stat.field = "Disturbance Cause Category Type",
             max.num = 20,
             export.plot = TRUE)

# pie chart on Overall Condition State Type, counts
d <- ref_hps(db.con = db.con,
             d = d,
             stat = "pie",
             stat.name = "Overall Condition State Type",
             stat.field = "Overall Condition State Type",
             export.plot = TRUE)

# pie chart on Overall Condition State Type, percents
d <- ref_hps(db.con = db.con,
             d = d,
             stat = "pie",
             stat.name = "Overall Condition State Type",
             stat.field = "Overall Condition State Type",
             perc = TRUE,
             rounded = 1,
             export.plot = TRUE,
             dirOut = "C:/Rprojects/eamenaR/results/",
             fig.width = 12,
             fig.height = 6)

# ...
d <- hash::hash()

}
