% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_hps.R
\name{ref_hps}
\alias{ref_hps}
\title{Statistics about EAMENA Heritage places (spatial distribution)}
\usage{
ref_hps(
  db.con = NA,
  d = NA,
  stat.name = "eamena-hps",
  stat.format = ".geojson",
  plot.map = F,
  export.map = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  date.after = NA,
  date.before = Sys.Date(),
  verbose = TRUE
)
}
\arguments{
\item{db.con}{the parameters for the Postgresql EAMENA DB, in a `RPostgres::dbConnect()` format.}

\item{d}{a hash() object (a Python-like dictionary).}

\item{stat.name}{the name of the output file. By default "eamena-hps".}

\item{stat.format}{the extension of the geographic file (".geojson", ".shp"). GeoJSON by default.}

\item{plot.map}{if TRUE will plot a map (FALSE by default).}

\item{export.map}{if TRUE will export the map (FALSE by default).}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export.map is TRUE.}

\item{date.after}{the date after which the calculation is made. Usefull to limit the analysis. By default, NA.}

\item{date.before}{the date before which the calculation is made. Usefull to limit the analysis. By default, the current date (`Sys.Date()`)}

\item{verbose}{if TRUE (by default), print messages}
}
\value{
a hash() object. If plot.map and export.map are set to TRUE will also create and save maps
}
\description{
statistics about EAMENA Heritage places
}
\examples{

d <- hash::hash()
my_con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                               user = 'xxx',
                               password = 'xxx',
                               dbname = 'eamena',
                               host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                               port = 5432)
d <- ref_hps(db.con = my_con,
            d = d,
            date.after = '2021-12-31',
            date.before = '2023-01-01',
            stat.name = "eamena-hps-2022",
            stat.format = ".shp",
            dirOut = 'C:/Rprojects/eamena-arches-dev/data/geojson/',
            export.map = TRUE)

}
