% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_ids.R
\name{ref_ids}
\alias{ref_ids}
\title{Find the equivalences between eamenaR concepts and specific Arches instances}
\usage{
ref_ids(
  r.name = NA,
  ids.path = paste0(system.file(package = "eamenaR"), "/extdata/ids.csv")
)
}
\arguments{
\item{r.name}{the eamenaR label, by default NA.}

\item{ids.path}{the path to the correspondance table.}
}
\value{
A string which is the label used in the DB
}
\description{
This function read a table of correspondences between labels used in this R package, and labels used of a given instance of Arches. For example 'id' in this eamenaR package refers the 'EAMENA ID' in the EAMENA DB. This is useful to generalise the function to Arches instances having different labels. See the table here: https://github.com/eamena-oxford/eamenaR/blob/main/inst/extdata/ids.csv
}
\examples{

ref_ids("id")
# "EAMENA ID"

}
