% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_periods.R
\name{ref_periods}
\alias{ref_periods}
\title{Basic statistics on GeoJSON file like lists or charts}
\usage{
ref_periods(
  stat = c("order"),
  ref.periods =
    "https://raw.githubusercontent.com/achp-project/cultural-heritage/main/periodo-projects/cultural_periods.tsv",
  verbose = TRUE
)
}
\arguments{
\item{stat}{type of statistic that will be computed. Default `order` (ordered/seriated periods on their mean).}

\item{ref.periods}{the periods reference table.}

\item{verbose}{if TRUE (by default), print messages.}

\item{name}{of field on which paths will be grouped. For example "route". Will create as many plots as there are different categories. Default NA.}
}
\value{
Depending on the parameter `stat`. If `stat` = `order` will return a vector of period labels.
}
\description{
Workout on EAMENA cultural periods
}
\examples{

# return the ordered vector of cultural periods
ref_periods()

}
