% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_routes.R
\name{ref_routes}
\alias{ref_routes}
\title{Join the GeoJSON file of heritage places (nodes) with the CSV file of routes (edges)}
\usage{
ref_routes(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"), "/extdata/caravanserail_paths.csv"),
  concept.name = "hp.id",
  by = NA,
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded. By default 'caravanserail_paths.csv'.}

\item{concept.name}{the concept that will be retrieve from the `ids.csv` file. By default "hp.id".}

\item{by}{the field in the CSV file}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{
A `sf` file of the heritage places with their routes.
}
\description{
Join the GeoJSON file of heritage places with the CSV file of routes. Each heritage place has its route (the `by` option) registered in a CSV file. This function creates a merge of these two files
}
