% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_users.R
\name{ref_users}
\alias{ref_users}
\title{Statistics about Arches project users}
\usage{
ref_users(
  db.con = NA,
  d = NA,
  db.name = "EAMENA",
  stat = c("all"),
  stat.type = c("all"),
  stat.name = "users",
  plot.g = F,
  export.plot.g = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  date.after = NA,
  date.before = Sys.Date(),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)
}
\arguments{
\item{db.con}{the parameters for the Postgresql EAMENA DB, in a `RPostgres::dbConnect()` format.}

\item{d}{a hash() object (a Python-like dictionary).}

\item{db.name}{the name of the Arches database, by default `EAMENA`.}

\item{stat}{the type of statistic that will be computed. This is also the the hash dictionnary (`d`) field name that will be filled with this statistics, e.g. "total_users", "date_joined", etc., or "all". By default: "all".}

\item{stat.type}{the type of stat chart, or diagram that will be plotted. Choice: "edtf" for cumulative function, etc. By default "all".}

\item{stat.name}{the name of the output file. By default "users".}

\item{plot.g}{if TRUE will create a plot (FALSE by default).}

\item{export.plot.g}{if TRUE will export the plot (FALSE by default).}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export.plot.g is TRUE.}

\item{date.after}{the date after which the calculation is made. Usefull to limit the analysis. By default, NA.}

\item{date.before}{the date before which the calculation is made. Usefull to limit the analysis. By default, the current date (`Sys.Date()`)}

\item{fig.width, fig.height}{dimension of the exported plot in cm.}

\item{verbose}{if TRUE (by default), print messages}
}
\value{
a hash() object. If plot.g and export.plot.g are set to TRUE will also create and save plots
}
\description{
statistics about EAMENA users (total number, activities, etc.)
}
\examples{

d <- hash::hash()
my_con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                               user = 'xxx',
                               password = 'xxx',
                               dbname = 'eamena',
                               host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                               port = 5432)
d <- ref_users(db.con = my_con,
               d = d,
               date.after = "2020-08-01",
               plot.g = T,
               export.plot.g = T,
               fig.width = 14)

}
