% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_from_eamenaid.R
\name{uuid_from_eamenaid}
\alias{uuid_from_eamenaid}
\title{Return the UUID of a HP from EAMENA ID}
\usage{
uuid_from_eamenaid(
  db.name = "eamena",
  db.user = "postgres",
  db.password = NA,
  db.host = NA,
  db.port = 5432,
  d,
  eamenaid,
  field.uuid = "uuid",
  field.eamenaid = "eamenaid"
)
}
\arguments{
\item{db.name}{the name of the database, by default 'eamena'.}

\item{db.user}{the name of the user, by default 'postgres'.}

\item{db.password}{the user database password.}

\item{db.host}{the name, or the IP, of the database.}

\item{db.port}{the database port number, by default: 5432.}

\item{d}{a hash() object (a Python-like dictionary).}

\item{eamenaid}{a EAMENA ID (eg. "EAMENA-0187363").}

\item{field.uuid}{the name of the field that will be created in the a hash() object for the UUID. By default 'uuid'.}

\item{field.eamenaid}{the name of the field that will be created in the a hash() object for the EAMENA ID. By default 'eamenaid'.}
}
\value{
a hash() object (a Python-like dictionary) with EAMENA ID and UUID.
}
\description{
Return the UUID of a HP from EAMENA id and store it into a
hash() object alongside the EAMENA id. A connection with the database is needed
}
\examples{
d_sql <- hash::hash()
d_sql <- uuid_from_eamenaid(db.name = 'eamena',
                            db.password = 'postgis',
                            db.host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                            d = d_sql,
                            eamenaid = "EAMENA-0187363")
d_sql$uuid
# [1] "12053a2b-9127-47a4-990f-7f5279cd89da"

}
