% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_addZ.R
\name{geojson_addZ}
\alias{geojson_addZ}
\title{Add a Z value to geometries}
\usage{
geojson_addZ(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  geojson.out = NA,
  elevation.api = "open-elevation",
  timeout = 30,
  sleep = 0.3,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  verbose = T
)
}
\arguments{
\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{geojson.out}{the name of the output file. By default NA, will add the suffix 'Z' to the the name of the GeoJSON file (variable `geojson.path`).}

\item{elevation.api}{the geoserver API that will be used to collect the elevation, by default 'open-elevation' (https://api.open-elevation.com/). Another option is 'gmrt' (https://www.gmrt.org).}

\item{timeout}{the threshold below which the expectation of a response is accepted, in seconds. By default, 30.}

\item{sleep}{the time delay between two API request in seconds. By default 0.3.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export plot is TRUE.}

\item{verbose}{if TRUE (by default): verbose.}
}
\value{

}
\description{
Use a geoserver API to recover the Z for a given set of points stored in a GeoJSON file. The geoserver API can lead to a Timeout.
}
\examples{

df <- geojson_addZ(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
                   dirOut = "C:/Rprojects/eamenaR/inst/extdata/")


}
