% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_csv.R
\name{geojson_csv}
\alias{geojson_csv}
\alias{geojson_kml}
\title{Convert from GeoJSON to CSV}
\usage{
geojson_kml(
  geom.path = paste0(system.file(package = "eamenaR"), "/extdata/Waypoints.kmz"),
  geom.types = c("POINT", "LINE", "POLYGON"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  geojson.name = "Waypoints",
  selectedName = "EAMENA.ID",
  selectedFields = c("EAMENA.ID", "Resource.Name", "resourceid"),
  verbose = T
)
}
\arguments{
\item{geom.path}{the path to the GeoJSON file. This file comes from the geojson_kml() function.}

\item{export}{if TRUE (by default), will export the GeoJSON to a CSV file.}

\item{dirOut}{the path to the folder where the CSV file will be created.}

\item{verbose}{if TRUE (by default) then display different messages.}

\item{csv.name}{the name of the CSV that will be created without the extension}
}
\value{
a CSV file
}
\description{
Convert from KML/KMZ to GeoJSON, or from GeoJSON to KML. Geometries drawn in Google Earth or Google Earth pro can be exported as KML or KMZ (ie, compressed KML). At the same time, geometries drawn in EAMENA can be exported as GeoJSON. For a given HP, this data can be then imported into the EAMENA DB.
}
\examples{

}
