% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_csv.R
\name{geojson_csv}
\alias{geojson_csv}
\title{Convert from GeoJSON to CSV}
\usage{
geojson_csv(
  geom.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_outGeoJSON.geojson"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  csv.name = "caravanserail_outCSV",
  verbose = T
)
}
\arguments{
\item{geom.path}{path to GeoJSON file. File comes from the `geojson_kml()` function.}

\item{export}{if TRUE (default), will export GeoJSON to CSV file.}

\item{dirOut}{path to folder where CSV file will be created.}

\item{csv.name}{name of CSV that will be created without the extension}

\item{verbose}{if TRUE (default) then display different messages.}
}
\value{
a CSV file
}
\description{
Function allows file format conversion from GeoJSON to CSV. Output CSV file can be imported into EAMENA DB with bulk upload append procedure.
}
\examples{

geojson_csv()

}
