% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_format_path.R
\name{geojson_format_path}
\alias{geojson_format_path}
\title{Format data on paths}
\usage{
geojson_format_path(geojson.path = geojson.path., csv.path = csv.path.)
}
\arguments{
\item{geojson.path}{the path of the GeoJSON file.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded.}
}
\value{
A dataframe with the appropriate format
}
\description{
Use a dataframe of heritage places (like vertices) inherited from a GeoJSON, and a file of paths (like edges) inherited form a CSV, between these heritage places, to format a new dataframe. This new dataframe is used by thefunctions `geojson_map_path()`, `geojson_boxplot_path()`. The function cleans the datasets removing existing paths liking two heritage places taht not exist in the GeoJSON
}
\examples{

}
