% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_format_path.R
\name{geojson_format_path}
\alias{geojson_format_path}
\title{Format data on paths}
\usage{
geojson_format_path(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  by.category = "route",
  concept.name = "hp.id",
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded. By default 'caravanserail_paths.csv'.}

\item{by.category}{the name of the category. By default "route" for caravanserais.}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{
A dataframe with the appropriate columns: "from.id", "from", "to.id", "to", "from.geom", "to.geom", "path.wkt", "dist.m", "route"
}
\description{
Use a dataframe of heritage places (ie, places, vertices), and a file of paths (as edges) between these heritage places, to format a new dataframe. The heritage places (HP) are stored in a GeoJSON file. The paths between these HP are stored in a CSV file. The function cleans the datasets removing existing paths linking two heritage places that don't not exist in the GeoJSON. This new dataframe is used by the functions `geojson_map_path()`, `geojson_boxplot_path()` to model the dataset with network analysis.
}
\examples{

}
