% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_format_path.R
\name{geojson_format_path}
\alias{geojson_format_path}
\title{Alignment of heritage place data and heritage place paths data}
\usage{
geojson_format_path(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"), "/extdata/caravanserail_paths.csv"),
  by = NA,
  concept.name = "hp.id",
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{path of GeoJSON file. Default 'caravanserail.geojson'.}

\item{csv.path}{path to CSV where edges between two heritage places are recorded. Default 'caravanserail_paths.csv'.}

\item{concept.name}{concept that will be retrieve from the `ids.csv` file. Default "hp.id".}

\item{verbose}{if TRUE (default), print messages.}

\item{name}{of field on which paths will be grouped. Example "route". Will create as many plots as there are different categories. Default NA.}
}
\value{
Dataframe with appropriate columns: "from.id", "from", "to.id", "to", "from.geom", "to.geom", "path.wkt", "dist.m", "route"
}
\description{
Use dataframe of heritage places (ie, places, vertices), and file of paths (as edges) between these heritage places, to format new dataframe. Heritage places (HP) are stored in GeoJSON file. Paths between HP are stored in CSV file. Function cleans the datasets, removing existing paths linking two heritage places that don't not exist in the GeoJSON. New dataframe is used by functions `geojson_map_path()`, `geojson_boxplot_path()` to model dataset with network analysis.
}
