% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_get_field.R
\name{geojson_get_field}
\alias{geojson_get_field}
\title{Get values of a given field in a GeoJSON file.}
\usage{
geojson_get_field(geojson.path, field = "EAMENA.ID")
}
\arguments{
\item{geojson.path}{the path to the GeoJSON file,
eg: "https://raw.githubusercontent.com/eamena-oxford/eamena-arches-dev/main/data/geojson/caravanserail.geojson"}

\item{field}{a field name, in R format, eg: EAMENA.ID}
}
\value{
A vector with all values
}
\description{
Get values of a given field
}
\examples{

geojson_get_field(geojson.path = "https://raw.githubusercontent.com/eamena-oxford/eamena-arches-dev/main/data/geojson/caravanserail.geojson")

}
