% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_kml.R
\name{geojson_kml}
\alias{geojson_kml}
\title{Convert from/to GeoJSON, KMZ or KML}
\usage{
geojson_kml(
  geom.path = paste0(system.file(package = "eamenaR"), "/extdata/Waypoints.kmz"),
  geom.types = c("POINT", "LINE", "POLYGON"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  geojson.name = "Waypoints",
  verbose = T
)
}
\arguments{
\item{geom.path}{the path to the KML, KMZ or GeoJSON file.}

\item{geom.types}{the types of geometries ("POINT", "LINE" or "POLYGON") that will be selected. By default all: `c("POINT", "LINE", "POLYGON")`.}

\item{export}{if TRUE, will export the KML/KMZ file in a GeoJSON format, or the GeoJSON file as a KML, if FALSE simple plot.}

\item{dirOut}{the path to the folder where the KML/KMZ/GeoJSON file will be created.}

\item{geojson.name}{the name of the KML/KMZ/GeoJSON that will be created without the extension}

\item{verbose}{if TRUE (by default) then display different messages.}
}
\value{
a GeoJSON file or a KML file
}
\description{
Convert from KML/KMZ to GeoJSON, or from GeoJSON to KML. Geometries drawn in Google Earth or Google Earth pro can be exported as KML or KMZ (ie, compressed KML). At the same time, geometries drawn in EAMENA can be exported as GeoJSON. For a given HP, this data can be then imported into the EAMENA DB.
}
\examples{

library(dplyr)

# from KMZ to GeoJSON
geom_kml(geom.path = "C:/Rprojects/eamenaR/inst/extdata/Waypoints.kmz",
         dirOut = "C:/Rprojects/eamenaR/inst/extdata/",
         geom.types = "POLYGON",
         export = T,
         geojson.name = "Waypoints_outGeoJSON")

# from GeoJSON to KML
geom_kml(geom.path = "C:/Rprojects/eamenaR/inst/extdata/kites.geojson",
         dirOut = "C:/Rprojects/eamenaR/inst/extdata/",
         export = T,
         geojson.name = "kites_outKML")

}
