% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_map.R
\name{geojson_map}
\alias{geojson_map}
\title{Create a map, whether static or interactive leaflet, from a GeoJSON file}
\usage{
geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  ids = "EAMENA.ID",
  field.names = NA,
  highlights.ids = NA,
  stamen.zoom = 8,
  plotly.plot = F,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8
)
}
\arguments{
\item{map.name}{the name of the output map. By default "map".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{ids}{the IDs of the resources, by default "EAMENA.ID" (R fieldname format, without spaces)}

\item{field.names}{a vector one or many field names for thematic cartography. If NA (by default)
will create a general map}

\item{highlights.ids}{EAMENA IDs (ex: 'EAMENA-0205783') that will be highlighted in the map.
If NA (by default), no highlights}

\item{stamen.zoom}{the zoom of the Stamen basemap, between 0 (world, unprecise) to 18 (building, very precise)}

\item{plotly.plot}{if FALSE create a static PNG, if TRUE create a plotly plot as a HTML widget}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export plot is TRUE}
}
\value{
A map interactive (leaflet) or not
}
\description{
Create a distribution map
}
\examples{

# plot a general map of heritage places
 geojson_map(map.name = "caravanserail")

# save different thematic map
geojson_map(map.name = "caravanserail",
           field.names = c("Disturbance.Cause.Type.", "Damage.Extent.Type"),
           export.plot = T)

# plot a general map of geoarchaeological data
geojson_map(map.name = "geoarch",
            ids = "GEOARCH.ID",
            stamen.zoom = 6,
            geojson.path = "C:/Rprojects/eamena-arches-dev/data/geojson/geoarchaeo.geojson",
            export.plot = F)

}
