% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_map.R
\name{geojson_map}
\alias{geojson_map}
\title{Create a map, whether static or interactive leaflet, from a GeoJSON file}
\usage{
geojson_map(
  map.name = "map",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  highlights.eamenaids = NA,
  plotly.plot = F,
  export.plot = F,
  dataOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8
)
}
\arguments{
\item{map.name}{the name of the output map. By default "map".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{highlights.eamenaids}{EAMENA IDs (ex: 'EAMENA-0205783') that will be highlighted in the map.
If NA (by default), no highlights}

\item{plotly.plot}{if FALSE create a static PNG, if TRUE create a plotly plot as a HTML widget}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it}

\item{dataOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export plot is TRUE}
}
\value{
A map interactive (leaflet) or not
}
\description{
Create a distribution map
}
\examples{

}
