% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_map_path.R
\name{geojson_map_path}
\alias{geojson_map_path}
\title{Create a map with paths between different heritage places}
\usage{
geojson_map_path(
  map.name = "map_path",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  stamen.zoom = 8,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8
)
}
\arguments{
\item{map.name}{the name of the output map and the name of the saved file (if export.plot is TRUE). By default "map_path".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded.}

\item{stamen.zoom}{the zoom of the Stamen basemap, between 0 (world, unprecise) to 21 (building, very precise). By default NA, the zoom level will be calculated automatically.}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export plot is TRUE.}
}
\value{
A PNG map of heritage places linked together by paths
}
\description{
Create a distribution map of heritage places linked together by paths, for example for caravanserails
}
\examples{

# plot a general map of heritage places
 geojson_map(map.name = "caravanserail")

# save a map
geojson_map_path(map.name = "map_paths_4",
                 geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_2.geojson",
                 csv.path = "C:/Rprojects/eamena-arches-dev/projects/caravanserail/caravanserail_paths_3.csv",
                 export.plot = T,
                 fig.width = 14,
                 fig.height = 11,
                 dirOut = "C:/Rprojects/eamena-arches-dev/projects/caravanserail/")

}
