% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_map_path.R
\name{geojson_map_path}
\alias{geojson_map_path}
\title{Create a map and a profile with paths between different heritage places}
\usage{
geojson_map_path(
  map.name = "map_path",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  csv.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail_paths.csv"),
  export.type = c("map"),
  by = "route",
  selected.category = NA,
  interactive = FALSE,
  stamen.zoom = NA,
  export.plot = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  color.set = "Set1",
  verbose = TRUE
)
}
\arguments{
\item{map.name}{the name of the output map and the name of the saved file (if export.plot is TRUE). By default "map_path".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'.}

\item{csv.path}{the path to the CSV where the edges between two heritage places are recorded. By default 'caravanserail_paths.csv'.}

\item{export.type}{the type of output: a map (`map`) or a profile (`profile`). For this latter the Z should be calculated with the `geojson_addZ()` function.}

\item{by}{the name of the category column. By default "route" for caravanserais.}

\item{selected.category}{limit the study to some categories. For example to some particular routes for caravanserais. By default NA, no limitation.}

\item{interactive}{if TRUE will plot a VisNetwork. By default FALSE.}

\item{stamen.zoom}{the zoom of the Stamen basemap, between 0 (world, unprecise) to 21 (building, very precise). By default NA, the zoom level will be calculated automatically.}

\item{export.plot}{if TRUE, export the plot, if FALSE will only display it.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export plot is TRUE.}

\item{fig.width, fig.height}{size of the output map.}

\item{color.set}{the RBrewer color set. By default "Set1".}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{
A PNG map and an HMTL map of heritage places linked together by paths
}
\description{
Create a distribution map and an elevation profile of heritage places linked together by paths, for example for caravanserails.
}
\examples{

library(dplyr)

# plot a general map of heritage places
geojson_map_path(map.name = "caravanserail_paths", export.plot = F)

# plot and export
geojson_map_path(map.name = "caravanserail_paths", export.plot = T, fig.width = 11)

# create an interactive map of each route
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T)

# create an interactive map of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 interactive = T,
                 export.plot = T,
                 dirOut = "C:/Rprojects/eamenaR/results/")

# create the profile of each route
df <- geojson_addZ(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
                   dirOut = "C:/Rprojects/eamenaR/inst/extdata/")
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 export.type = "profile")

# create the profile of each route and export
geojson_map_path(geojson.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserailZ.geojson",
                 csv.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail_paths.csv",
                 selected.category = c(0, 1, 2, 3, 4),
                 export.type = "profile",
                 export.plot = T,
                 fig.height = 11,
                 fig.width = 18,
                 dirOut = "C:/Rprojects/eamenaR/results/")

}
