% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_measurements.R
\name{geojson_measurements}
\alias{geojson_measurements}
\title{Measurements on a GeoJSON file}
\usage{
geojson_measurements(
  stat.name = "stat",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  ids = "EAMENA.ID",
  stat = c("area"),
  plot.stat = T,
  export.stat = F,
  write.stat = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fig.width = 8,
  fig.height = 8,
  verbose = TRUE
)
}
\arguments{
\item{stat.name}{the name of the output file. By default "stat".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{ids}{the IDs of the resources, by default "EAMENA.ID" (n.b: R fieldname format, without spaces)}

\item{stat}{the statistic that will be computed. By default 'area'.}

\item{plot.stat}{if TRUE (by default) will plot the stat as a graphic}

\item{export.stat}{if TRUE return the stats to be stored in a new variable}

\item{write.stat}{if TRUE, export the stats in a new file.}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is write.stat is TRUE}

\item{verbose}{if TRUE (by default) verbose}
}
\value{
Show or export measurements statistics on the GeoJSOn file
}
\description{
Compute measurements (areas, L x l, etc.) on a GeoJSON file
}
\examples{

geojson_measurements(stat.name = "areas", export.stat = T)

}
