% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_shp.R
\name{geojson_shp}
\alias{geojson_shp}
\title{Convert between GeoJSON and SHP}
\usage{
geojson_shp(
  geom.path = paste0(system.file(package = "eamenaR"), "/extdata/Waypoints.kmz"),
  geom.types = c("POINT", "LINE", "POLYGON"),
  export = T,
  dirOut = paste0(system.file(package = "eamenaR"), "/extdata/"),
  geojson.name = "Waypoints",
  select.name = "EAMENA ID",
  select.fields = c("EAMENA ID", "Resource Name", "resourceid"),
  verbose = T
)
}
\arguments{
\item{geom.path}{the path to the SHP/GeoJSONfile.}

\item{geom.types}{the types of geometries ("POINT", "LINE" or "POLYGON") that will be selected. By default all: `c("POINT", "LINE", "POLYGON")`.}

\item{export}{if TRUE, will export the SHP file in a GeoJSON format, or the GeoJSON file as a SHP, if FALSE simple plot.}

\item{dirOut}{the path to the folder where the SHP/GeoJSON file will be created.}

\item{geojson.name}{the name of the SHP/GeoJSON that will be created without the extension}

\item{select.name}{for SHP export only. The field selected to be the SHP name of the HP, by default `"EAMENA ID"`.}

\item{select.fields}{for SHP export only. SHP conversion remove a large number of GeoJSON fields. This variable is used to select the fields we want to preserve. By default: `c("EAMENA ID","Resource Name", "resourceid")`.}

\item{verbose}{if TRUE (by default) then display different messages.}
}
\value{
a GeoJSON file or a SHP file, depending on the input file.
}
\description{
Convert from SHP (shapefile) to GeoJSON, or from GeoJSON to SHP.
}
\examples{

library(dplyr)

# from GeoJSON to SHP
geojson_shp(geom.path = "C:/Rprojects/eamenaR/inst/extdata/caravanserail.geojson",
            geom.types = "POLYGON",
            dirOut = "C:/Rprojects/eamenaR/results/",
            export = T,
            geojson.name = "caravanserailOUT")

}
