% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_stat.R
\name{geojson_stat}
\alias{geojson_stat}
\title{Basic statistics on a GeoJSON file}
\usage{
geojson_stat(
  stat.name = "stat",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  ids = "EAMENA.ID",
  stat = c("list_fields"),
  export.stat = F,
  write.stat = F,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{stat.name}{the name of the output file. By default "stat".}

\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{ids}{the IDs of the resources, by default "EAMENA.ID" (n.b: R fieldname format, without spaces)}

\item{stat}{the statistic that will be computed. By default 'list_fields' (list the fields)
The other options are: "list_ids" list EAMENA IDs ; etc.}

\item{export.stat}{if TRUE return the stats to be stored in a new variable}

\item{write.stat}{if TRUE, export the stats in a new file, if FALSE will only display it}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is write.stat is TRUE}
}
\value{
Show or export basic statistics on the GeoJSON file
}
\description{
Basic descriptive statistics on a GeoJSON file
}
\examples{

geojson_stat(stat.name = "geojson_fields", export.stat = T)

}
