% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_stat.R
\name{geojson_stat}
\alias{geojson_stat}
\title{Basic statistics on a GeoJSON file}
\usage{
geojson_stat(
  stat.name = "stat",
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  stat = c("list_fields"),
  export.stat = F,
  dataOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{geojson.path}{the path of the GeoJSON file. By default 'caravanserail.geojson'}

\item{stat}{the statistic that will be computed. By default 'list_fields' (list the fields)}

\item{export.stat}{if TRUE, export the stats, if FALSE will only display it}

\item{dataOut}{the folder where the outputs will be saved. By default: '/results'.
If it doesn't exist, it will be created. Only useful is export plot is TRUE}
}
\value{
A map interactive (leaflet) or not
}
\description{
Create a distribution map
}
\examples{

}
