% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bbox.R
\name{geom_bbox}
\alias{geom_bbox}
\title{Create the bounding box of BUs' geometries and save it as a GeoJSON file}
\usage{
geom_bbox(
  dataDir = NA,
  x_column = "xcoord",
  y_column = "ycoord",
  wkt_column = NA,
  dirOut = dataDir,
  geojson.name = "geojson.geojson"
)
}
\arguments{
\item{dataDir}{the path to the folder where the XLSXs are}

\item{x_column, y_column}{the column of the X and Y coordinates, if NA the function will look the field 'wkt'.}

\item{wkt_column}{the column of the WKT coordinates. Useful if the X,Y coordinates are already in}

\item{dirOut}{the path to the folder where the GeoJSON file will be created.}

\item{geojson.name}{the name of the GeoJSON that will be created.}
}
\value{
a GeoJSON file
}
\description{
Read one or various XLSX worksheets within a folder. Collect the xmin, xmax, ymin, xmax and creates the bounding box, that is to say, the extent of the geometries. Can be used to preselect grids, to get stamen (ggmap), etc. Used before the function \link[eamenaR]{geom_within_gs} and \link[eamenaR]{list_mapping_bu.R} to get the IDs of the potential grids.
}
\examples{

geom_bbox(dataDir = "C:/Rprojects/eamena-arches-dev/data/bulk/bu/")

}
