% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_within_gs.R
\name{geom_within_gs}
\alias{geom_within_gs}
\title{Test if a resource (HP) geometry is within a Grid Square (gs).}
\usage{
geom_within_gs(
  resource.wkt = NA,
  gs.path = paste0(system.file(package = "eamenaR"), "/extdata/grid_squares.geojson"),
  verbose = TRUE
)
}
\arguments{
\item{resource.wkt}{the WKT geometry of a resource, as a character format. This WKT geometry can comes from a BU sheet (ex: "POINT(0.9 35.8)").}

\item{gs.path}{a path to a GeoJSON file, by default to the example 'grid_squares.geojson' This GeoJSON is an EAMENA output of the Grids as a GeoJSON URL. The name of the grids can be calculated with the \link[eamenaR]{geom_bbox} function. The GeoJSON is read and convert to a `sf` object.}

\item{verbose}{if TRUE (by default): verbose.}
}
\value{
the ID of the Grid Square for each WKT geometries.
}
\description{
Test if the geometry of a resource (eg. Heritage Place) is within a Grid Square (gs). If so, return the ID of the Grid Square.
}
\details{
when this function is called for a dataframe, it has to be nested into a loop.
}
\examples{

library(dplyr)

geom_within_gs(resource.wkt = "POINT(0.9 35.8)")


}
