% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_within_gs.R
\name{geom_within_gs}
\alias{geom_within_gs}
\title{Test if resource geometry is within a Grid Square (gs).}
\usage{
geom_within_gs(
  resource.wkt = NA,
  gs.path = paste0(system.file(package = "eamenaR"), "/extdata/grid_squares.geojson"),
  verbose = TRUE
)
}
\arguments{
\item{resource.wkt}{WKT geometry of resource, as character format. This WKT geometry can come from a bulk upload sheet (ex: "POINT(0.9 35.8)").}

\item{gs.path}{path to GeoJSON file, default to the example 'grid_squares.geojson' This GeoJSON is an EAMENA output of Grids as a GeoJSON URL. Name of grids can be calculated with \link[eamenaR]{geom_bbox} function. GeoJSON is read and convert to `sf` object.}

\item{verbose}{if TRUE (default): verbose.}
}
\value{
the ID of Grid Square for each WKT geometries.
}
\description{
Test if geometry of a resource, example heritage place (HP) is within a Grid Square (gs). If so, return ID of Grid Square.
}
\details{
This function must be nested in a loop when called for a database.
}
\examples{

library(dplyr)

geom_within_gs(resource.wkt = "POINT(0.9 35.8)")


}
