% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_edtf.R
\name{plot_edtf}
\alias{plot_edtf}
\title{Create an aoristic analysis of threats from a dataset}
\usage{
plot_edtf(
  data_file = paste0(system.file(package = "eamenaR"),
    "/extdata/disturbances_edtf.xlsx"),
  date_column = "EDTF",
  site_column = "S_ID",
  cause_column = "Disturbance.Cause",
  type_column = "Disturbance.Type",
  edft_limits = "2004-01-01..2019-12-31",
  edtf_span = "myd",
  edtf_analyse = c("all", "category"),
  edtf_analyse_category = "Disturbance.Type",
  edft_round = 4,
  id.filter = NA,
  type.plot = c("plotly"),
  export.plot = FALSE,
  file_out = "df_syria_out2",
  dirOut = paste0(system.file(package = "eamenaR"), "/results/")
)
}
\arguments{
\item{data_file}{the path to the file where the XLSX is}

\item{date_column}{the column of the EDTF data}

\item{site_column}{the column of the site names}

\item{cause_column}{the column of the Causes}

\item{type_column}{the column of the Type of threats}

\item{edtf_span}{the time span of the analysis: "myd" = years, months and days (by default);
"my" = years and months, etc.}

\item{edtf_analyse}{type of analysis. If "all" will sum the different categories,
if "category" will plot all categories of threats}

\item{edtf_analyse_category}{the field name of the threats category that will be analysed.
"Disturbance.Type" (by default), or "Disturbance.Cause".}

\item{edft_round}{round density. By default 4 decimals}

\item{id.filter}{if not NA, will filter on this subset of sites (ex: id.filter = c("AM009"))}

\item{type.plot}{file extension of the plot to export. Either "plotly" for an HTML dynamic plot,
or "png" for a static one.}

\item{export.plot}{if TRUE will export, if FALSE (by default) will show}

\item{edtf_limits}{the min and max of the temporal anaylsis. Intersect with the limits of the dataset}

\item{wkt_column}{the column of the WKT coordinates. Useful if the X,Y coordinates are already in}

\item{dataOut}{the path to the folder where the GeoJSON file will be created.}

\item{geojson.name}{the name of the GeoJSON that will be created}
}
\value{
a plot
}
\description{
Use an XLSX file where R doesn't accept spaces in headers, so the spaces are replaced by dots.
}
\examples{

library(dplyr)
plot_edtf(data_file = "C:/Rprojects/eamena-arches-dev/data/time/Disturbances_EDTF.xlsx")

}
