% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_are_duplicates.R
\name{ref_are_duplicates}
\alias{ref_are_duplicates}
\title{Help to identify if two EAMENA HP are real duplicates or not}
\usage{
ref_are_duplicates(
  db.con = NA,
  d = NA,
  field = "are_duplicates",
  resourceid.list = c("563567f7-eef0-4683-9e88-5e4be2452f80",
    "fb0a2ef4-023f-4d13-b931-132799bb7a6c"),
  selected.fields = c("Assessment.Investigator...Actor", "Assessment.Activity.Date",
    "Resource.Name"),
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/caravanserail.geojson"),
  round.dist = 2,
  export.table = FALSE,
  dirOut = paste0(system.file(package = "eamenaR"), "/results/"),
  fileOut = "duplicates.xlsx",
  verbose = TRUE
)
}
\arguments{
\item{db.con}{the parameters for the Postgresql EAMENA DB, in a RPostgres::dbConnect() format. If null, will use a GeoJSON file.}

\item{d}{a hash() object (a Python-like dictionary).}

\item{field}{the field name that will be created in the a hash() object.}

\item{resourceid.list}{a list with the ResourceID to compare. By default: c("563567f7-eef0-4683-9e88-5e4be2452f80", "fb0a2ef4-023f-4d13-b931-132799bb7a6c")}

\item{selected.fields}{the list of fields that will be selected to compare their values between different potential duplicates. By default the GeoJSON geometry is also selected.}

\item{round.dist}{an integer for the number of digit to preserve in the distance computing. By default: 2.}

\item{export.table}{if TRUE will export the table of duplicates (FALSE by default).}

\item{dirOut}{the folder where the outputs will be saved. By default: '/results'. If it doesn't exist, it will be created. Only useful is export.plot.g is TRUE.}

\item{fileOut}{the output file name. It could be an XLSX or a CSV file. Only useful is export.plot.g is TRUE. By default "duplicates.xlsx".}

\item{verbose}{if TRUE (by default), print messages.}
}
\value{
a matrix stored in hash() object. This matrix has the ResourceID of the compared HP in column.  If export.table is set to TRUE it will also create an CSV or XLSX table with the potential duplicates, and the fuzzy matching value (column 'dist')
}
\description{
values coming from different records will be put one against another to facilitate the comparisons between possible duplicates. A fuzzy matching between these values is computed and stored in the colum 'dist' to resume the information. The higher this value, the closer are the compared values.
}
\examples{

d <- hash::hash()
d <- ref_are_duplicates(d = d,
                        export.table = T,
                        fileOut = "duplicates.csv")
}
