% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_eamenaid.R
\name{uuid_eamenaid}
\alias{uuid_eamenaid}
\title{Return the UUID of a HP from EAMENA ID, or the opposite}
\usage{
uuid_eamenaid(
  db.con = NA,
  d = NA,
  id = NA,
  field.eamenaid = "eamenaid",
  field.resourceid = "resourceid",
  disconn = TRUE
)
}
\arguments{
\item{d}{a hash() object (a Python-like dictionary).}

\item{id}{a EAMENA ID (eg. "EAMENA-0187363") or a ResourceID (eg. "12053a2b-9127-47a4-990f-7f5279cd89da").}

\item{disconn}{if TRUE (default) will disconnect from the DB once done. If FALSE, the user has to disconnect (eg. DBI::dbDisconnect(my_con)).}

\item{db.name}{the name of the database, by default 'eamena'.}

\item{db.user}{the name of the user, by default 'postgres'.}

\item{db.password}{the user database password.}

\item{db.host}{the name, or the IP, of the database.}

\item{db.port}{the database port number, by default: 5432.}

\item{field.uuid}{the name of the field that will be created in the a hash() object for the UUID. By default 'uuid'.}

\item{field.id}{the name of the field that will be created in the a hash() object for the EAMENA ID. By default 'id'.}
}
\value{
a hash() object (a Python-like dictionary) with EAMENA ID and ResourceID. If a given ID doesn't exist, will fill the value of the hash dictionary with NA.
}
\description{
Return the ResourceID of a HP from its EAMENA ID, or the opposite: the ResourceID from the EAMENA ID, and store these ID into a hash() object. A connection with the EAMENA database is needed. The ResourceID is a UUID
}
\examples{

d <- hash::hash()
my_con <- RPostgres::dbConnect(drv = RPostgres::Postgres(),
                               user = 'postgres',
                               password = 'xxx',
                               dbname = 'xxx',
                               host = 'ec2-54-155-109-226.eu-west-1.compute.amazonaws.com',
                               port = 5432)

# from the EAMENA ID
d <- uuid_eamenaid(db.con = my_con,
                   d = d,
                   id = "EAMENA-0187363",
                   disconn = FALSE)
d$resourceid
# [1] "12053a2b-9127-47a4-990f-7f5279cd89da"

# from the ResourceID
d <- uuid_eamenaid(db.con = my_con,
                   d = d,
                   id = "12053a2b-9127-47a4-990f-7f5279cd89da")
d$eamenaid
# [1] "EAMENA-0187363"

}
